+++ src/postgres/query-datum.ts
@@ -149,7 +149,7 @@ class PostgresQueryDatumPartial<T = any> extends AbstractDatumPartial<T> impleme
           else
             throw new Error('Cannot use "div" without something selected!');
           break;
-        case 'div':
+        case 'mod':
           if(sel2)
             throw new Error('Can only use "eq" and "ne" on sub-object!');
           else if(sel)